+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.0] - [Unrelease]
+
+### Added
+
+#### Supabase
+
+- add source column logs table migration
+
 ## [0.3.0] - 2026-02-24
 
 ### Added